Revert "Project dir restructure"
