+++ auth.py
@@ -1,14 +1,10 @@
-import threading
+from kivymd.uix.snackbar import Snackbar
 from kivy.clock import Clock
-
-from app.commons.toast import Toaster
-from app.utils.colors import Colors
-from app.commons.loader import Loading
-import config.config as config
+from loading import Loading
+import config
+import threading
 from kivymd.app import MDApp
-from app.google_auth import GoogleOAuth
-
-
+from google_auth import GoogleOAuth
 
 
 class Authentication:
@@ -35,7 +31,7 @@ class Authentication:
         """Method to call to start the login process."""
         self.loading.open()
         self.login_thread.start()
-    
+
     def after_login(self, token):
         """
         Is called after the login process is successful.
@@ -58,8 +54,7 @@ class Authentication:
             elif status_code == 401:
                 Clock.schedule_once(
                     lambda *args: (
-                        Toaster(message="You don't have access to this screen", bg_color=Colors().ErrorColor.get("BackgroundColor"),
-                    font_size=14).toast(),
+                        Snackbar(text="You don't have access to this screen").open(),
                     ),
                     0,
                 )
@@ -67,26 +62,18 @@ class Authentication:
         except requests.exceptions.RequestException:
             self.error_listener("Backend server is down")
 
-
-
-   
     def error_listener(self, msg):
         """Called whenever there is an error in the login process"""
 
         Clock.schedule_once(
             lambda *args: (
                 self.loading.dismiss(),
-                Toaster(message=msg, bg_color=Colors().ErrorColor.get("BackgroundColor"),
-                    font_size=14).toast(),
+                Snackbar(text=msg).open(),
             ),
             0,
         )
 
 
-    def login_and_register_user(self):
-        """Custom login response to hold user return responses"""
-        return Authentication().login()
-    
 def change_screen(root, resp):
     root.current = "dashboard"
     root.get_screen("dashboard").ids.dashboard.ids.welcome_text.text = resp.text
