+++ google_auth/globals.py
