--- kiv/screens/login_screen.kv
@@ -1,38 +1,0 @@
-<LoginScreenContent>:
-    md_bg_color: app.COLORS.WhiteColor.get("BackgroundColor")
-    BoxLayout:
-        orientation: "vertical"
-        spacing: dp(30)
-        padding: dp(41)
-        pos_hint: {"center_x": 0.5, "center_y": 0.5}
-        size_hint_x: None
-        width: dp(400)
-
-        MDLabel:
-            text: "TurntablRide"
-            color: 0, 0, 0
-            font_size: 35
-            size_hint_y: None
-            height: dp(10)
-            halign: "center"
-            font: "fn_italic"
-
-        Image:
-            source: "images/landing_big.png"
-            size_hint_y: 0.6
-            border: app.COLORS.WhiteColor.get("BackgroundColor")
-            background_normal: app.COLORS.WhiteColor.get("BackgroundColor")
-
-        MDFillRoundFlatIconButton:
-            id: login_with_google_btn
-            text: "  Login with Google"
-            icon: "images/google_icon.png"
-            icon_color: "orange"
-            radius: [20, 20, 20, 20]
-            md_bg_color: app.COLORS.BlueColor.get("BackgroundColor")
-            line_color: app.COLORS.BlueColor.get("TextColor")
-            text_color: app.COLORS.WhiteColor.get("BackgroundColor")
-            on_press: root.login_with_google()
-            size: dp(295), dp(48)
-            pos_hint: {"center_x": 0.5}
-            font: app.Fonts.fonts.get("jakarta_normal")
