+++ kv_files/landing.kv
@@ -1,0 +1,55 @@
+#:kivy 2.1.0
+#: import Authentication auth.Authentication
+
+
+<OurScreenManager>:
+    LoginScreen:
+    DashboardScreen:
+
+
+<LoginView@BoxLayout>:
+    orientation: "vertical"
+    padding: [0, 0, 0, 100]
+    Label:
+        text: "turntablRide"
+        color: 0, 0, 0
+        bold: True
+        font_size: 35
+        size_hint: 1.0, 0.2
+    Image:
+        source: "images/landing_big.png"
+    BoxLayout:
+        orientation: "horizontal"
+        size_hint: 0.3, 0.1
+        pos_hint: {"x": 0.35}
+        Image:
+            source: "images/google_icon.png"
+            size_hint: 0.3, 1.0
+        Button:
+            text: "Login with Google"
+            bold: True
+            font_size: 18
+            background_color: 0.234, 0.59, 0.965, 1
+            background_normal: ""
+            on_press: Authentication().login()
+
+<DashBoardView@BoxLayout>:
+    orientation: "vertical"
+    Label:
+        id: welcome_text
+        text: "Welcome Back"
+        color: 0, 0, 0
+        bold: True
+        size_hint: 0.3, 0.12
+        pos_hint: {"x": 0.35}
+
+
+<DashboardScreen@Screen>:
+    name: "dashboard"
+    DashBoardView:
+        id: dashboard
+
+
+<LoginScreen@Screen>:
+    name: "login"
+    LoginView:
