+++ kv_files/loading.kv
@@ -1,7 +1,7 @@
 #:kivy 2.1.0
 #:import Factory kivy.factory.Factory
 
-<Loader>:
+<Loading>:
     title: ""
     size_hint_x: None
     size_hint_y: None
