--- run.kv
@@ -1,22 +1,0 @@
-#:import LoginScreenContent app.screens.login_screen
-## Include reusable components to the top
-# You should import and add your screen content after declaring your screen <MyWidget@BaseClass>:
-#     for example 
-#     `` #:import LoginScreenContent app.screens.login_screen ``
-#     `` LoginScreenContent `` is the class name which contain the contents of the login_screen
-#     and it should be declared under the LoginScreen class (actual screen for the login 
-#     created inside the run.py file). Finally register the LoginScreen under WindowManager:
-#       WindowManager:
-#         LoginScreen: (from .run.py file)
-#   and then tag it 
-#     as shown below 
-#       <LoginScreen>:
-#           name: "_login_screen" (just an tag name of screen)
-#           LoginScreenContent (from the app/screens/login_screen )
-
-WindowManager:
-    LoginScreen:
-
-<LoginScreen>:
-    name: "_login_screen"
-    LoginScreenContent
